Moved functions returning fort files lengths to utils
