+++ tests/utils.py
@@ -1,10 +1,12 @@
 import inspect
+import os
 from typing import (Any,
                     Callable,
                     Iterable,
+                    Iterator,
                     Hashable,
                     Dict,
-                    List)
+                    List,)
 
 from hypothesis import (Verbosity,
                         find,
@@ -38,3 +40,26 @@ def sub_dict(dictionary: Dict[Hashable, Any],
              keys: Iterable[Hashable]) -> Dict[Hashable, Any]:
     return {key: dictionary[key]
             for key in keys}
+
+
+def fort_files_by_metallicities_lengths(fort_links: Dict[int, Iterator],
+                                        *,
+                                        base_dir: str) -> List[int]:
+    lengths = []
+    for metallicity, fort_links_range in fort_links.items():
+        lengths.extend(fort_files_lengths(fort_links=fort_links_range,
+                                          base_dir=base_dir))
+    return lengths
+
+
+def fort_files_lengths(fort_links: Iterator,
+                       *,
+                       base_dir: str) -> List[int]:
+    lengths = []
+    fort_link_dirs = [os.path.join(base_dir,
+                                   'fort_files/fort.' + str(fort_link))
+                      for fort_link in fort_links]
+    for fort_link_dir in fort_link_dirs:
+        lengths.append(sum(1 for line in open(fort_link_dir)))
+
+    return lengths
